EcoEnchants Informational Menu Update!
I've updated the menu to accommodate for the new enchantments added in by the plugin developer, they're all listed under the correct categories.
Along with that, with the addition of a new curse, it made the curse category menu uneven, and unpleasing, so I added an explanation to what "Curses" are in the menu to make it even, and pleasing. thanks, discord is listed on repo if any support is needed.